requirements.txt
alabaster==0.7.12
Babel==2.8.0
boto3==1.12.26
botocore==1.15.26
certifi==2018.11.29
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
Django==2.2.13
django-colorfield==0.2.2
django-cors-headers==3.3.0
django-extensions==2.2.8
django-widget-tweaks==1.4.5
docutils==0.15.2
idna==2.8
imagesize==1.2.0
Jinja2==2.11.2
jmespath==0.9.5
jwcrypto==0.7
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
packaging==20.4
pem==20.1.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.10.1
python-jwt==3.2.6
pytz==2018.9
requests==2.21.0
s3transfer==0.3.3
six==1.14.0
snowballstemmer==2.0.0
Sphinx==3.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.0
urllib3==1.24.2
xlrd==1.2.0
